refactor: split SearchStreamImpl into three focused classes
Extract document-to-entity deserialization into SearchStreamDocumentMapper
and query-building/execution logic into SearchStreamQueryExecutor, reducing
SearchStreamImpl from 1336 to ~620 lines with no behaviour changes.
